requirements.txt
numpy==1.24.3
torch==2.0.1
Pillow==10.0.0
matplotlib==3.7.2
tqdm==4.65.0